requirements.txt
Django==6.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.5.1
python-dotenv==1.0.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
